Ranking funcional, ao terminar jogo, perdendo ou ganhando jogador escolhe se continua a jogar

+++ src/showdomilhao/models/RankItem.java
@@ -11,5 +11,5 @@ package showdomilhao.models;
  */
 public class RankItem {
     public String jogador;
-    public int pontuacao;
+    public long pontuacao;
 }
